requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudscraper==1.2.69
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.4
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.4
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
